Move app/utils.js into app/utils folder
