+++ src/components/ModelTableList/OwnerGroup.js
@@ -1,6 +1,9 @@
 import { useSelector } from "react-redux";
 import MainTable from "@canonical/react-components/dist/components/MainTable";
-import { generateStatusElement, getModelStatusGroupData } from "app/utils";
+import {
+  generateStatusElement,
+  getModelStatusGroupData,
+} from "app/utils/utils";
 import { getGroupedByOwnerAndFilteredModelData } from "app/selectors";
 import { generateModelDetailsLink, getStatusValue } from "./shared";
